feat(modal): add antd modal method and useModal
#551
